+++ contracts/interfaces/IPlayer.sol
@@ -2,9 +2,9 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 interface IPlayer {
-    function joinGame(uint256 gameId, uint256 card) external payable;
+    function joinGame(uint64 gameId, uint256 card) external payable;
 
-    function openCard(uint256 gameId,string key, uint8 content) external;
+    function openCard(uint64 gameId,string key, uint8 content) external;
 
-    function timeoutGame(uint256 gameId) external;
+    function timeoutGame(uint64 gameId) external;
 }
